# RONDB - 338: Optimise REST Server