For benchmarking REST batchpkread,

From
```log
    performance_test.go:149: -------------------------------------------------
    performance_test.go:143: Number of requests:         7843
    performance_test.go:144: Batch size (per requests):  100
    performance_test.go:145: Number of threads:          7
    performance_test.go:146: Throughput:                 79417.880637 pk lookups/second
    performance_test.go:147: 50th percentile latency:    8 ms
    performance_test.go:148: 99th percentile latency:    35 ms
    performance_test.go:149: -------------------------------------------------
BenchmarkSimple-8           9530           1462199 ns/op         2460698 B/op      13516 allocs/op
```

to:
```log
    performance_test.go:143: Number of requests:         8016
    performance_test.go:144: Batch size (per requests):  100
    performance_test.go:145: Number of threads:          7
    performance_test.go:146: Throughput:                 92915.990723 pk lookups/second
    performance_test.go:147: 50th percentile latency:    7 ms
    performance_test.go:148: 99th percentile latency:    29 ms
    performance_test.go:149: -------------------------------------------------
BenchmarkSimple-8          11149           1320736 ns/op         2147834 B/op      10115 allocs/op